requirements.txt
Django==5.2
django_imagekit==5.0.0
django_model_utils==5.0.0
djangorestframework_simplejwt==5.5.0
drf_yasg==1.21.10
Faker==37.4.0
imagekitio==4.1.0
Pillow==11.3.0
python-dotenv==1.1.1
Requests==2.32.4
twilio==9.6.3
django-cors-headers==4.7.0
django_otp==1.6.1
django-filter
# PhonePe SDK for latest payment integration
--index-url https://phonepe.mycloudrepo.io/public/repositories/phonepe-pg-sdk-python
--extra-index-url https://pypi.org/simple
phonepe_sdk
